% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate,character,character,logical-method}
\alias{annotate,character,character,logical-method}
\title{Annotate probesets with a chip type}
\usage{
\S4method{annotate}{character,character,logical}(object, target, check.target)
}
\arguments{
\item{object}{A vector of chacater strings}

\item{target}{An chiptype in \code{ribiosAnnotation::gtiChiptypes()}}

\item{check.target}{Logical, whether checking the target exists or not}
}
\description{
Annotate probesets with a chip type
}
