% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate,character,character,missing-method}
\alias{annotate,character,character,missing-method}
\title{Annotate probesets with a chip type without checking}
\usage{
\S4method{annotate}{character,character,missing}(object, target)
}
\arguments{
\item{object}{A vector of chacater strings}

\item{target}{An chiptype in \code{ribiosAnnotation::gtiChiptypes()}}
}
\description{
Annotate probesets with a chip type without checking
}
