% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate,eSet,character,logical-method}
\alias{annotate,eSet,character,logical-method}
\title{Annotate an eSet object with a chip type}
\usage{
\S4method{annotate}{eSet,character,logical}(object, target, check.target)
}
\arguments{
\item{object}{An eSet}

\item{target}{An chiptype in \code{ribiosAnnotation::gtiChiptypes()}}
}
\description{
Annotate an eSet object with a chip type
}
