% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate,eSet,character,missing-method}
\alias{annotate,eSet,character,missing-method}
\title{Annotate an eSet object with a chip type, do not perform checking}
\usage{
\S4method{annotate}{eSet,character,missing}(object, target)
}
\arguments{
\item{object}{An eSet}

\item{target}{An chiptype in \code{ribiosAnnotation::gtiChiptypes()}}
}
\description{
Annotate an eSet object with a chip type, do not perform checking
}
