% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate,ExpressionSet,character,logical-method}
\alias{annotate,ExpressionSet,character,logical-method}
\title{Annotate an ExpressionSet object with a chip type}
\usage{
\S4method{annotate}{ExpressionSet,character,logical}(object, target, check.target)
}
\arguments{
\item{object}{An ExpressionSet}

\item{target}{An chiptype in \code{ribiosAnnotation::gtiChiptypes()}}

\item{check.target}{Logical, whether checking the target exists or not}
}
\description{
Annotate an ExpressionSet object with a chip type
}
