% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate,ExpressionSet,character,missing-method}
\alias{annotate,ExpressionSet,character,missing-method}
\title{Annotate an ExpressionSet object with a chip type, do not perform checking}
\usage{
\S4method{annotate}{ExpressionSet,character,missing}(object, target)
}
\arguments{
\item{object}{An ExpressionSet}

\item{target}{An chiptype in \code{ribiosAnnotation::gtiChiptypes()}}
}
\description{
Annotate an ExpressionSet object with a chip type, do not perform checking
}
