% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFullRank.R
\name{assertFullRank}
\alias{assertFullRank}
\title{Assert whether a matrix is of full rank numerically}
\usage{
assertFullRank(matrix)
}
\arguments{
\item{matrix}{Numeric matrix}
}
\value{
If not full rank, the function stops. Otherwise, an invisible \code{TRUE} is returned
}
\description{
Assert whether a matrix is of full rank numerically
}
\examples{

myMat <- matrix(c(1,1,1,0,1,1), ncol=2, byrow=FALSE)
assertFullRank(myMat)
}
