% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ChipFetcher2ExpressionSet}
\alias{ChipFetcher2ExpressionSet}
\title{Import ChipFetcher export files into an ExpressionSet object.}
\usage{
ChipFetcher2ExpressionSet(filename, chip, orthologue = FALSE)
}
\arguments{
\item{filename}{Exported file name of ChipFetcher}

\item{chip}{Chip type used to annotate the features, e.g.
\code{HG-U133_PLUS_2}. In case missing, chips are automatically mapped.
Assigning the value of \code{chip} accelerates the probe mapping step.}

\item{orthologue}{Logical, whether features should be mapped to human
orthologues? By default \code{FALSE}.}
}
\value{
An \code{ExpressionSet} object.
}
\description{
Import files exported by Roche web-tool \code{ChipFetcher} into an
\code{ExpressionSet} object.
}
\details{
\code{chip} and \code{orthologue} are only valid when
\code{ribiosAnnotation} is available.
}
\references{
\url{http://bioinfo.bas.roche.com:8080/bicgi/chipfetcher}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
