% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{contrastAnnotation<-}
\alias{contrastAnnotation<-}
\title{Assign contrast annotation to an object}
\usage{
contrastAnnotation(object) <- value
}
\arguments{
\item{object}{An object, see supported methods below}

\item{value}{Contrast annotation data.frame}
}
\value{
The modified object
}
\description{
Assign contrast annotation to an object
}
