% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{contrastAnnotation}
\alias{contrastAnnotation}
\alias{contrastAnnotation,DesignContrast-method}
\alias{contrastAnnotation<-,DesignContrast-method}
\title{Extract the contrast annotation data.frame from an object}
\usage{
contrastAnnotation(object)

\S4method{contrastAnnotation}{DesignContrast}(object)

\S4method{contrastAnnotation}{DesignContrast}(object) <- value
}
\arguments{
\item{object}{An object, see supported methods below}

\item{value}{A contrast annotation data.frame}
}
\value{
A data.frame annotating contrasts
}
\description{
Extract the contrast annotation data.frame from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{DesignContrast}: Return the contrast annotation data.frame
from a DesignContrast object

\item \code{DesignContrast}: Assign a contrast annotation data.frame
to a \code{contrastContrast} object
}}

