% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{contrastMatrix,MArrayLM-method}
\alias{contrastMatrix,MArrayLM-method}
\title{Extract contrast matrix from MArrayLM}
\usage{
\S4method{contrastMatrix}{MArrayLM}(object)
}
\arguments{
\item{object}{A MArrayLM object from the limma package}
}
\value{
contrast matrix
}
\description{
Extract contrast matrix from MArrayLM
}
