% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{contrastMatrix<-}
\alias{contrastMatrix<-}
\title{Assign contrast matrix to an object}
\usage{
contrastMatrix(object) <- value
}
\arguments{
\item{object}{An object, see supported methods below}

\item{value}{Contrast matrix}
}
\value{
The modified object
}
\description{
Assign contrast matrix to an object
}
