% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastNames,DesignContrast-method}
\alias{contrastNames,DesignContrast-method}
\title{Return contrast names of a DesignContrast object}
\usage{
\S4method{contrastNames}{DesignContrast}(object)
}
\arguments{
\item{object}{A DesignContrast object}
}
\description{
Return contrast names of a DesignContrast object
}
