% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{contrastNames,MArrayLM-method}
\alias{contrastNames,MArrayLM-method}
\title{Extract contrast names from MArrayLM}
\usage{
\S4method{contrastNames}{MArrayLM}(object)
}
\arguments{
\item{object}{A MArrayLM object from the limma package}
}
\value{
Character vector of contrast names
}
\description{
Extract contrast names from MArrayLM
}
