% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{contrastNames}
\alias{contrastNames}
\alias{contrastNames,DesignContrast-method}
\alias{contrastNames,MArrayLM-method}
\title{Extract contrastNames from an object}
\usage{
contrastNames(object)

\S4method{contrastNames}{DesignContrast}(object)

\S4method{contrastNames}{MArrayLM}(object)
}
\arguments{
\item{object}{An object, see supported methods below}
}
\value{
A character vector of contrast names
}
\description{
Extract contrastNames from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{contrastNames(DesignContrast)}: Return contrast names, i.e., column names of the
contrast matrix

\item \code{contrastNames(MArrayLM)}: Extract contrast names from an object of MArrayLM

}}
