% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{dataFrameTwoVecs}
\alias{dataFrameTwoVecs}
\title{Build a data.frame from two vectors of potential different lengths}
\usage{
dataFrameTwoVecs(vec1, vec2, col.names = c("Vec1", "Vec2"))
}
\arguments{
\item{vec1}{A vector}

\item{vec2}{Another vector}

\item{col.names}{A character vector of length 2 giving column names of the output \code{data.frame}.

The shorter vector of the two are extended to the same length by appending empty strings.}
}
\value{
A \code{data.frame} of two columns. The row count matches the longer vector
}
\description{
Build a data.frame from two vectors of potential different lengths
}
