% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{design2group}
\alias{design2group}
\title{Infer groups from a design matrix}
\usage{
design2group(designMatrix)
}
\arguments{
\item{designMatrix}{A design matrix}
}
\value{
A factor vector giving the groups inferred from the design matrix

A naive logic is used: samples of the same design vectors are of the same group.

The inference is known to fail when control variables, such as age or RIN numbers, 
vary between samples of the same group.
}
\description{
Infer groups from a design matrix
}
\examples{
myDesign <- model.matrix(~gl(3,3))
design2group(myDesign)
}
