% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DesignContrast-class}
\alias{DesignContrast-class}
\alias{show,DesignContrast-method}
\title{Study design and contrast information}
\description{
The DesignContrast class represents key information in a designed experiment
}
\section{Objects from the Class}{
 Objects can be created by calls of the
function \code{DesignContrast}. However, the users should not directly call this
function. Instead, \code{\link{parseDesignContrast}} should be called.
}

\keyword{classes}
