% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{DesignContrast}
\alias{DesignContrast}
\title{Contrast a DesignContrast object}
\usage{
DesignContrast(
  designMatrix,
  contrastMatrix = NULL,
  groups = NULL,
  dispLevels = NULL
)
}
\arguments{
\item{designMatrix}{A design matrix}

\item{contrastMatrix}{A contrast matrix. If null, no comparison can be done.}

\item{dispLevels}{A character vector of the same length as the number of levels encoded by \code{group},
indicating how different groups should be labelled. If missing, levels of \code{group} are used.}

\item{group}{A factor vector of the same length as the number of columns of the design matrix. 
If missing, \code{design2group} is used to infer groups.}
}
\value{
A DesignContrast object
}
\description{
Contrast a DesignContrast object
}
\examples{
myFac <- gl(3,3, labels=c("baseline", "treat1", "treat2"))
myDesign <- model.matrix(~myFac)
colnames(myDesign) <- c("baseline", "treat1", "treat2")
myContrast <- limma::makeContrasts(contrasts=c("treat1", "treat2"), levels=myDesign)
DesignContrast(myDesign, myContrast, groups=myFac)
DesignContrast(myDesign, myContrast, groups=myFac, dispLevels=c("C", "T1", "T2"))
}
