% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{designMatrix,MArrayLM-method}
\alias{designMatrix,MArrayLM-method}
\title{Extract design matrix from MArrayLM}
\usage{
\S4method{designMatrix}{MArrayLM}(object)
}
\arguments{
\item{object}{A MArrayLM object from the limma package}
}
\value{
Design matrix
}
\description{
Extract design matrix from MArrayLM
}
