% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{designMatrix<-}
\alias{designMatrix<-}
\title{Assign design matrix to an object}
\usage{
designMatrix(object) <- value
}
\arguments{
\item{object}{An object, see supported methods below}

\item{value}{Design matrix}
}
\value{
The modified object
}
\description{
Assign design matrix to an object
}
