% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{designVariables}
\alias{designVariables}
\alias{designVariables,DesignContrast-method}
\title{Extract design variable names from an object}
\usage{
designVariables(object)

\S4method{designVariables}{DesignContrast}(object)
}
\arguments{
\item{object}{An object, see supported methods below}
}
\value{
A character vector of variable names
}
\description{
Extract design variable names from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{DesignContrast}: Return the names of variables (column names) 
in the design matrix of a DesignContrast object
}}

