% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{exprsToLong,matrix-method}
\alias{exprsToLong,matrix-method}
\title{Transform an expression matrix to long table}
\usage{
\S4method{exprsToLong}{matrix}(
  x,
  idvar = "illID",
  timevar = "hybridID",
  valuevar = "value",
  ids = rownames(x),
  valueType = "raw"
)
}
\arguments{
\item{x}{A matrix or an ExpressionSet object}

\item{idvar}{Variable name of the feature identifier, passed to \code{\link{reshape}}}

\item{timevar}{The time variable, passed to \code{\link{reshape}}}

\item{valuevar}{The value variable}

\item{ids}{Feature identifiers}

\item{valueType}{Character string, value type}
}
\value{
A \code{data.frame}
}
\description{
Transform an expression matrix to long table
}
