% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{formatGmt}
\alias{formatGmt}
\alias{formatGmt,character,character,character-method}
\alias{formatGmt,character,missing,character-method}
\alias{formatGmt,character,character,list-method}
\alias{formatGmt,character,missing,list-method}
\title{Make strings in the GMT format}
\usage{
formatGmt(title, comment, genes)

\S4method{formatGmt}{character,character,character}(title, comment, genes)

\S4method{formatGmt}{character,missing,character}(title, genes)

\S4method{formatGmt}{character,character,list}(title, comment, genes)

\S4method{formatGmt}{character,missing,list}(title, genes)
}
\arguments{
\item{title}{Character, title(s) of gene set(s)}

\item{comment}{Character, comment(s) of gene set(s). Can be of the same
length as the \code{title}, or be of length one: in the latter case, it will
be replicated in gene set. This option can also be left out: the comment
field of the GMT file will be left blank.}

\item{genes}{A character vector of gene names, or a list of such vectors. In
the former case, one GMT line is produced; otherwise multiple lines are
returned. In the latter case, the length of the list must match the length
of \code{title}.}
}
\value{
One or more lines of GMT file
}
\description{
Resulting string(s) can be exported into \code{GMT} file by
\code{\link{writeLines}}
}
\section{Methods (by class)}{
\itemize{
\item \code{title = character,comment = character,genes = character}: title, comment, and genes are one character string

\item \code{title = character,comment = missing,genes = character}: title and genes are both one character string, comments are missing

\item \code{title = character,comment = character,genes = list}: title and comments are both vectors of character 
strings, genes are a list of the same length

\item \code{title = character,comment = missing,genes = list}: title is vectors of character strings, comments are 
missing, genes are a list of the same length as the title
}}

\examples{

formatGmt(title="GeneSet0", comment="My geneset", genes=c("MAPT", "MAPK", "AKT1"))
formatGmt(title="GeneSet0", genes=c("MAPT", "MAPK", "AKT1"))

formatGmt(title=c("GeneSet0", "GeneSet1"),
          comment=c("My geneset 0", "My geneset 1"),
          genes=list(c("MAPT", "MAPK", "AKT1"), c("EGFR", "CDC42")))
formatGmt(title=c("GeneSet0", "GeneSet1"),
          comment="My genesets",
          genes=list(c("MAPT", "MAPK", "AKT1"), c("EGFR", "CDC42")))
formatGmt(title=c("GeneSet0", "GeneSet1"),
          genes=list(c("MAPT", "MAPK", "AKT1"), c("EGFR", "CDC42")))
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
