% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{groups}
\alias{groups}
\alias{groups,DesignContrast-method}
\title{Extract sample groups from an object}
\usage{
groups(object)

\S4method{groups}{DesignContrast}(object)
}
\arguments{
\item{object}{An object, see supported methods below}
}
\value{
A factor of sample groups
}
\description{
Extract sample groups from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{groups(DesignContrast)}: Return the raw sample groups from a DesignContrast object

}}
