% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{isInputDesignConsistent}
\alias{isInputDesignConsistent}
\title{Test whether the input design matrix is consistent with the sample names}
\usage{
isInputDesignConsistent(descon, sampleNames)
}
\arguments{
\item{descon}{A DesignContrast object}

\item{sampleNames}{A vector of string characters, specifying sample names

If the sample names in DesignContrast are identical with the given sample names,
an invisible \code{TRUE} is returned.

If the two sets are identical, however the order of sample names do not match, a warning
message is raised, and an invisible \code{FALSE} is returned

If the two sets have differences, the mismatching sample names are printed for diagnosis.}
}
\value{
A invisible logical value. \code{TRUE} if and only if the sample names match perfectly.
}
\description{
Test whether the input design matrix is consistent with the sample names
}
