% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDgeTable.R
\name{limmaDgeTable}
\alias{limmaDgeTable}
\title{Return dgeTable from a marrayLM object}
\usage{
limmaDgeTable(marrayLM, contrast = NULL, confint = TRUE)
}
\arguments{
\item{marrayLM}{An object returned by ‘lmFit’ and ‘eBayes’}

\item{contrast}{NULL, or a character string indicating the contrast of
interest}

\item{confint}{Logical, whether confidence intervals should be returned}
}
\value{
A \code{data.frame}
}
\description{
Return dgeTable from a marrayLM object
}
