% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateDgeTable.R
\name{limmaTopTable2dgeTable}
\alias{limmaTopTable2dgeTable}
\title{Transform limma::topTable results to a DGEtable}
\usage{
limmaTopTable2dgeTable(limmaTopTable)
}
\arguments{
\item{limmaTopTable:}{topTable returned by limma::topTable}
}
\value{
A data.frame known as DGEtable which has controlled column names
}
\description{
Transform limma::topTable results to a DGEtable
}
\examples{
library(limma)
example.sd <- 0.3*sqrt(4/rchisq(100,df=4))
example.y <- matrix(rnorm(100*6,sd=example.sd),100,6)
example.y[1:2,4:6] <- example.y[1:2,4:6] + 2
rownames(example.y) <- paste("Gene",1:100)
example.design <- cbind(Grp1=1,Grp2vs1=c(0,0,0,1,1,1))
example.fit <- lmFit(example.y,example.design)
example.fit <- eBayes(example.fit)
example.tt <- topTable(example.fit, coef=2)
example.dt <- limmaTopTable2dgeTable(example.tt)
head(example.dt)
}
