% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{matrixToLongTable}
\alias{matrixToLongTable}
\title{Transform a matrix to long table}
\usage{
matrixToLongTable(x, valueLabel = "value", rowLabel = "row", colLabel = "col")
}
\arguments{
\item{x}{A matrix}

\item{valueLabel}{Character string, the label of the value}

\item{rowLabel}{Character string, the name of the column holding the row names}

\item{colLabel}{Character string, the name of the column holding the column names}
}
\value{
A \code{data.frame}
}
\description{
Transform a matrix to long table
}
\examples{
myMatrix <- matrix(rnorm(24), nrow=4, dimnames=list(LETTERS[1:4], letters[1:6]))
matrixToLongTable(myMatrix)
}
