% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{parseContrastStr}
\alias{parseContrastStr}
\title{Parse contrast from strings}
\usage{
parseContrastStr(contrastStr)
}
\arguments{
\item{contrastStr}{A vector of character strings}
}
\value{
A contrast matrix
}
\description{
Parse contrast from strings
}
