% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{parseDesignContrast}
\alias{parseDesignContrast}
\title{Parse study design and asked questions encoded in design and contrast matrices or in one-way ANOVA designs}
\usage{
parseDesignContrast(designFile = NULL, contrastFile = NULL,
  sampleGroups = NULL, groupLevels = NULL, dispLevels = NULL,
  contrasts = NULL)
}
\arguments{
\item{designFile:}{A plain tab-delimited file with headers encoding the design matrix, or NULL}

\item{contrastFile:}{A plain tab-delimited file with headers encoding the contrast matrix, or NULL}

\item{sampleGroups:}{A character string concatenated by commas (e.g. A,B,C), or a plain text file containing one string per line (e.g. A\emph{newline}B\emph{newline}C), encoding sample group memberships.}

\item{groupLevels:}{Similar format as 'sampleGroups', encoding levels (e.g. order) of the sampleGroups}

\item{dispLevels:}{Similar format as 'sampleGroups', encoding the display of the groupLevels. Must match 'groupLevels'}

\item{contrasts:}{Similar format as 'sampleGroups', encoding contrasts in case of one-way ANOVA designs}
}
\value{
A S4-object 'DesignContrast'
}
\description{
Parse study design and asked questions encoded in design and contrast matrices or in one-way ANOVA designs
}
\examples{
## one-way ANOVA
parseDesignContrast(sampleGroups="As,Be,As,Be,As,Be",groupLevels="Be,As", dispLevels="Beryllium,Arsenic", contrasts="As-Be")
## design/contrast matrix
designFile <- system.file("extdata/example-designMatrix.txt", package="ribiosExpression")
contrastFile <- system.file("extdata/example-contrastMatrix.txt", package="ribiosExpression")
# minimal information
parseDesignContrast(designFile=designFile, contrastFile=contrastFile)
# with extra information about sample groups
parseDesignContrast(designFile=designFile, contrastFile=contrastFile,sampleGroups="As,Be,As,Be,As,Be",groupLevels="Be,As", dispLevels="Beryllium,Arsenic")
}

