% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{parseDesignContrastFile}
\alias{parseDesignContrastFile}
\title{Parse design and contrast from files}
\usage{
parseDesignContrastFile(
  designFile,
  contrastFile,
  groupsStr = NULL,
  levelStr = NULL,
  dispLevelStr
)
}
\arguments{
\item{designFile}{A tab-delimited file encoding the design matrix}

\item{contrastFile}{A tab-delimited file encoding the contrast matrix}

\item{groupsStr}{A vector of character strings, giving sample groups}

\item{levelStr}{A vector of level strings}

\item{dispLevelStr}{A vector of strings to be used as display labels, if exist}
}
\value{
A DesignContrast object
}
\description{
Parse design and contrast from files
}
