% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{plot.DesignContrast}
\alias{plot.DesignContrast}
\title{Plot a DesignContrast object}
\usage{
\method{plot}{DesignContrast}(x, y = NULL, name = NULL, ...)
}
\arguments{
\item{x}{A \code{DesignContrast} object}

\item{y}{NULL, ignored}

\item{name}{Name of the object, used in the title of the heatmaps}

\item{...}{Other parameters passed to \code{\link{biosHeatmap}}}
}
\value{
An invisible list of two elements, containing return values of \code{\link{biosHeatmap}} for design and contrast matrices, respectively.
}
\description{
The function plots two heatmaps, one of the design matrix and the other of the contrast matrix.
}
