% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotationFile.R
\name{readAnnotationFile}
\alias{readAnnotationFile}
\title{Read in an annotation file in the tsv-format, with or without row names}
\usage{
readAnnotationFile(file, outputKeyName = "FeatureName", ...)
}
\arguments{
\item{file}{A tab-delimited file without quotes, the first column must
contain identifiers (key names). In case the first column has no column name 
and it contains row names, they will be used as feature names.}

\item{outputKeyName}{The key name used in the output \code{data.frame}. It 
does not need to exist in the column names of the input \code{data.frame}. 
In this situation, the first column will be used as output keys. If it does
exist in the column names of the input \code{data.frame}, the content of 
that column must be identical with the first column of the annotation file.}

\item{...}{Other parameters passed to \code{\link[ribiosIO]{readMatrix}}, 
which are further passed to \code{\link[utils]{read.table}}.}
}
\value{
A \code{data.frame} containing the annotation, with the first
  column named as \code{outputKeyName} that contains feature identifiers as character 
  strings. In case the input table contains the column with the same name, 
  the content in that column must match the row names, otherwise an error 
  is reported.
  
This function is called by \code{readFeatureAnnotationFile} and
 \code{readSampleAnnotationFile}. Normal users are unlikely to use it.
}
\description{
Read in an annotation file in the tsv-format, with or without row names
}
\examples{
f1 <- system.file("extdata",
  "featureAnnotation/featureAnnotationFile-withRowNames.txt", 
  package="ribiosExpression")
f2 <- system.file("extdata",
  "featureAnnotation/featureAnnotationFile-withoutRowNames.txt", 
  package="ribiosExpression")
# 'FeatureName' does not exist in the column names of f1
f1Read <- readAnnotationFile(f1, outputKeyName="FeatureName") 
# 'GeneID' exists in the colum names of f2, and it is the first column.
f2Read <- readAnnotationFile(f2, outputKeyName="GeneID")
head(f1Read)
head(f2Read)
}
