% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFeatureAnnotationFile.R
\name{readFeatureAnnotationFile}
\alias{readFeatureAnnotationFile}
\title{Read in feature annotation file in the tsv-format, with or without row names}
\usage{
readFeatureAnnotationFile(file, ...)
}
\arguments{
\item{file}{A tab-delimited file without quotes, the first column must
contain feauture identifiers. In case the first column has no column name 
and it contains row names, they will be used as feature names.}

\item{...}{Other parameters passed to \code{\link[ribiosIO]{readMatrix}}, 
which are further passed to \code{\link[utils]{read.table}}.}
}
\value{
A \code{data.frame} containing feature annotation, with the first
  column named as `FeatureName` that contains feature identifiers as character 
  strings. In case the input table contains the column `FeatureName`, the
  content in that column must match the row names, otherwise an error 
  is reported.
}
\description{
Read in feature annotation file in the tsv-format, with or without row names
}
\examples{
f1 <- system.file("extdata",
  "featureAnnotation/featureAnnotationFile-withRowNames.txt", 
  package="ribiosExpression")
f2 <- system.file("extdata",
  "featureAnnotation/featureAnnotationFile-withoutRowNames.txt", 
  package="ribiosExpression")
f1Read <- readFeatureAnnotationFile(f1)
f2Read <- readFeatureAnnotationFile(f2)
head(f1Read)
head(f2Read)
}
