% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_gct_cls.R
\name{readGct}
\alias{readGct}
\title{Import ExpressionSet into gct/cls files
the C version, about 5x faster than the R implementation for the ALL dataset}
\usage{
readGct(gct.file)
}
\arguments{
\item{gct.file}{Character string, the path to a GCT file}
}
\value{
A \code{ExpressionSet} object
}
\description{
Import ExpressionSet into gct/cls files
the C version, about 5x faster than the R implementation for the ALL dataset
}
