% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotationFile.R
\name{readSampleAnnotationFile}
\alias{readSampleAnnotationFile}
\title{Read in sample annotation file in the tsv-format, with or without row names}
\usage{
readSampleAnnotationFile(file, ...)
}
\arguments{
\item{file}{A tab-delimited file without quotes, the first column must
contain sample identifiers. In case the first column has no column name 
and it contains row names, they will be used as sample names.}

\item{...}{Other parameters passed to \code{\link[ribiosIO]{readMatrix}}, 
which are further passed to \code{\link[utils]{read.table}}.}
}
\value{
A \code{data.frame} containing sample annotation, with the first
  column named as `SampleName` that contains sample identifiers as character 
  strings. In case the input table contains the column `Samples`, the
  content in that column must match the row names, otherwise an error 
  is reported.
}
\description{
Read in sample annotation file in the tsv-format, with or without row names
}
\examples{
f1 <- system.file("extdata",
  "sampleAnnotation/sampleAnnotationFile-withRowNames.txt", 
  package="ribiosExpression")
f2 <- system.file("extdata",
  "sampleAnnotation/sampleAnnotationFile-withoutRowNames.txt", 
  package="ribiosExpression")
f1Read <- readSampleAnnotationFile(f1)
f2Read <- readSampleAnnotationFile(f2)
head(f1Read)
head(f2Read)
}
