% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{reannotate,ExpressionSet,logical-method}
\alias{reannotate,ExpressionSet,logical-method}
\title{Re-annotate an ExpressionSet with Roche Bioinformatics Infrastructure}
\usage{
\S4method{reannotate}{ExpressionSet,logical}(object, check.target)
}
\arguments{
\item{object}{An ExpressionSet object}

\item{check.target}{Logical, checking whether the annotation is supported}
}
\description{
Annotation type is inferred from the annotation label of Bioconductor
}
