% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{reannotate,ExpressionSet,missing-method}
\alias{reannotate,ExpressionSet,missing-method}
\title{Re-annotate an ExpressionSet with Roche Bioinformatics Infrastructure without checking}
\usage{
\S4method{reannotate}{ExpressionSet,missing}(object)
}
\arguments{
\item{object}{An ExpressionSet object}
}
\description{
Re-annotate an ExpressionSet with Roche Bioinformatics Infrastructure without checking
}
