% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\docType{methods}
\name{reannotate-methods}
\alias{reannotate-methods}
\alias{reannotate,eSet,logical-method}
\alias{reannotate,eSet,missing-method}
\title{Re-annotate: Transform an eSet object of Bioc-annotation to GTI-annotation}
\usage{
reannotate(object, check.target, ...)
}
\description{
The method accepts an object to re-annotate them. The action of
\sQuote{reannotating} defined in \code{ribiosExpression} is to transform an
eSet object of Bioconductor-annotation to GTI-annotation.
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"eSet\", check.target =
\"logical\")")}{ Currently the method only support an \code{eSet}
with a valid Bioconductor annotation set (for example \dQuote{hgu95av2}) in
the \code{annotation} slot. See \code{\link{reannotate}} for details and
examples.  } \item{list("signature(object = \"eSet\", check.target
=" \"missing\")")}{ \code{check.target} is set to FALSE when missing.
} }
}

\keyword{methods}
