% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/annotate.R
\docType{methods}
\name{reannotate}
\alias{reannotate}
\alias{reannotate,eSet,logical-method}
\alias{reannotate,eSet,missing-method}
\title{Transform an eSet object of Bioc-annotation into of GTI-annotation}
\usage{
reannotate(object, check.target, ...)

\S4method{reannotate}{eSet,logical}(object, check.target)

\S4method{reannotate}{eSet,missing}(object)
}
\arguments{
\item{object}{An \code{eSet} object, with the \code{annotation}
slot set as one of the valid annotations recognized by Bioconductor, for
instance \code{hgu95av2}.}

\item{check.target}{Logical, with \code{FALSE} as default. When set to
\code{TRUE}, beforing fetching database for annotations, the function first
checks whether the chip type is supported by GTI. If it is not the case, the
function will print error message and stop.}

\item{\dots}{Currently not implemented}
}
\value{
An \code{eSet} object with feature annotations updated by
GTI, and the \code{annotation} slot is changed to the chip type in GTI.
}
\description{
The function is used to transform an eSet object, which is
annotated by Bioconductor annotation packages, into an object with
annotation information from GTI.
}
\details{
The translation between Bioconductor annotation package names and GTI chip
types is performed by the \code{bioc2gti} function in the
\code{ribiosAnnotation} package.

Once the re-annotation succeeds, the \code{annotation} slot of the
\code{eSet} object will be overwritten by the corresponding chip
type name in GTI.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = eSet,check.target = logical}: Method for eSet and logical

\item \code{object = eSet,check.target = missing}: Method for eSet, not checking the target
}}

\examples{

data(ribios.ExpressionSet)
print(ribios.ExpressionSet)

\dontrun{
gti.eSet <- reannotate(ribios.ExpressionSet)
gti.eSet <- reannotate(ribios.ExpressionSet, check.target=FALSE)
print(gti.eSet)
}
}
\seealso{
\code{\link{annotate}} to annotate an \code{eSet} object
without prior information of bioc-annotation, or if that information is not
saved in the \code{annotation} slot.
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{methods}
