% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAllZeroVar.R
\name{removeAllZeroVar}
\alias{removeAllZeroVar}
\alias{removeAllZeroVar.matrix}
\alias{removeAllZeroVar.DesignContrast}
\title{Remove all-zero variables from design matrix and the corresponding contrast matrix}
\usage{
removeAllZeroVar(obj, contrasts)

\method{removeAllZeroVar}{matrix}(obj, contrasts)

\method{removeAllZeroVar}{DesignContrast}(obj, contrasts = NULL)
}
\arguments{
\item{obj}{Either a design matrix, rows are samples, columns are 
independent variables, and values are coefficients. Or a 
\code{DesignContrast} object that contains both design and contrast 
matrices. The design matrix may contain columns of all zeros, the 
independent variable corresponding to which can not be estimated.}

\item{contrasts}{Either \code{NULL}, in case \code{designMatrix} is a 
\code{DesignContrast} object, or a contrast matrix matching the design 
matrix, i.e. the number of rows must equal the number of column of the design matrix.
 Rows are independent variables, columns are names of the contrasts to be 
 estimated, the values are operations applied to the coefficients.}
}
\value{
Either a list of two matrices (\code{design} and 
  \code{contrasts}), or a \code{DesignContrast} object, 
  depending on the input parameter type. The design matrix and contrast 
  matrix have an attribute each, \code{notEstCoefs} and 
  \code{notEstContrasts}, that keep track of filtered variables and contrasts.
}
\description{
Remove all-zero variables from design matrix and the corresponding contrast matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{removeAllZeroVar(matrix)}: S3 function for matrix as input

\item \code{removeAllZeroVar(DesignContrast)}: S3 function for matrix as input

}}
\examples{
myTestDesign <- matrix(c(1,1,1,1, 1,1,0,0,0,0,1,1,0,0,0,0), 
  byrow=FALSE, nrow=4L, 
  dimnames=list(sprintf("S\%d", 1:4), c("Baseline", "Trt1", "Trt2", "Trt3")))
myTestContrast <- matrix(c(0,1,0,0, 0,0,1,0, 0,0,0,1), nrow=4L, byrow=FALSE,
  dimnames=list(colnames(myTestDesign), c("Trt1", "Trt2", "Trt3")))
removeAllZeroVar(myTestDesign, myTestContrast)
removeAllZeroVar(DesignContrast(myTestDesign, myTestContrast))
}
