% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosExpression-package.R
\docType{data}
\name{ribios.ExpressionSet}
\alias{ribios.ExpressionSet}
\title{An ExpressionSet for case demonstrations}
\format{
An \code{ExpressionSet} object.
}
\description{
This object is adapted from the \code{sample.ExpressionSet} object, with
feature annotations from GTI (Data stand: December 2011). It is used in case
studies where functionalities of the \code{ribiosExpression} package are
demonstrated.
}
\examples{

data(ribios.ExpressionSet)
tbl <- eSetToLongTable(ribios.ExpressionSet)
}
\references{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{datasets}
