% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniffFeatureType.R
\name{sniffFeatureType}
\alias{sniffFeatureType}
\title{Sniff the feature type of an object that implements the featureNames method}
\usage{
sniffFeatureType(object, majority = 0.5)
}
\arguments{
\item{object}{Any object that \code{featureNames} can be applied to, 
for instance an \code{ExpressionSet}, an \code{EdgeObject}, etc.}

\item{majority}{A numeric value, used for majority voting, passed to 
\code{guessFeatureType}.}
}
\value{
A character string indicating likely feature type.
}
\description{
Sniff the feature type of an object that implements the featureNames method
}
\seealso{
{\code{\link[ribiosAnnotation]{guessFeatureType}}}
}
