% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitPCA.R
\name{splitPCA}
\alias{splitPCA}
\title{Split an eSet object and run PCA on each split, return PCA scores as one data.frame}
\usage{
splitPCA(eset, factor, func = function(e) exprs(e), ...)
}
\arguments{
\item{eset}{An eSet object}

\item{factor}{One or more factor vectors, used to split the eSet object}

\item{func}{Function to retrieve values from split sub-eset objects}

\item{...}{Passed to \code{pcaScores}}
}
\value{
A \code{data.frame} of PCA scores combined from all splits.
}
\description{
Split an eSet object and run PCA on each split, return PCA scores as one data.frame
}
\examples{
data(ribios.ExpressionSet, package="ribiosExpression")
fac1 <- gl(2,13)
pcaScore1 <- splitPCA(ribios.ExpressionSet, fac1)
}
