% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateDgeTable.R
\name{truncateDgeTable}
\alias{truncateDgeTable}
\title{Truncate dgeTable into tables of positively and negatively differentially expressed genes according to the pre-defined criteria}
\usage{
truncateDgeTable(dgeTable)
}
\arguments{
\item{dgeTable}{dgeTable A DGEtable defined in ribiosExpression. Notice that the column names returned by limma::topTable are remapped (see limmaTopTable2dgeTable).}
}
\value{
A list of two elements: 'pos' and 'neg'. Each contains a dgeTable of positively/negatively regulated genes
}
\description{
Truncate dgeTable into tables of positively and negatively differentially expressed genes according to the pre-defined criteria
}
