% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_gct_cls.R
\name{writeCls}
\alias{writeCls}
\title{Write CLS (class file) from Eset}
\usage{
writeCls(eset, file = stdout(), sample.group.col = "group")
}
\arguments{
\item{eset}{An ExpressionSet object}

\item{file}{A file, by default the standard output}

\item{sample.group.col}{Column name in the sample annotation (\code{pData}) to be used as groups}
}
\description{
Write CLS (class file) from Eset
}
