% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosExpression-package.R
\docType{methods}
\name{writeGct-methods}
\alias{writeGct-methods}
\alias{writeGct,ExpressionSet-method}
\alias{writeGct,matrix-method}
\title{Write matrix or ExpressionSet in GCT file format}
\description{
Write matrix or ExpressionSet in GCT file format
}
\section{Methods}{
 \describe{ The S4-method is a wrapper of the
\code{write_gct} function implemented in the \code{ribiosIO} package.
\item{list("signature(obj = \"ExpressionSet\")")}{ An \code{ExpressionSet}
object. }

\item{list("signature(obj = \"matrix\")")}{ A matrix } }
}

\keyword{methods}
