% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVarMetadata.R
\name{writeVarMetadata}
\alias{writeVarMetadata}
\alias{writeVarMetadata.default}
\title{writexlxs: write AnnotatedDataFrame to a xlsx file}
\usage{
writeVarMetadata(x, path = tempfile(fileext = ".xlsx"), overwrite = TRUE)

\method{writeVarMetadata}{default}(x, path = tempfile(fileext = ".xlsx"), overwrite = TRUE)
}
\arguments{
\item{x}{An \code{ExpressionSet} object or an AnnotatedDataFrame.}

\item{path}{The xlsx file name to be written to.}

\item{overwrite}{Logical, whether the file should be overwritten if it exists.}
}
\value{
An invisible \code{TRUE} in case the file is successfully created, else \code{FALSE}.
}
\description{
writexlxs: write AnnotatedDataFrame to a xlsx file
}
\note{
We also tried the writexl package but the comments are not well supported by writexl, therefore we stay with openxlsx
}
\examples{
data("ribiosExpressionSet", package="ribiosExpression")
outfile <- tempfile()
writeVarMetadata(ribiosExpressionSet, path=outfile)
writeVarMetadata(phenoData(ribiosExpressionSet), path=outfile)
}
