% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{contrastAnnotation<-}
\alias{contrastAnnotation<-}
\title{Assign contrast annotaiton to an object}
\usage{
contrastAnnotation(object) <- value
}
\arguments{
\item{object}{An object, see supported methods below}

\item{value}{Contrast anaotation data.frame}
}
\description{
Assign contrast annotaiton to an object
}
