% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{contrastMatrix}
\alias{contrastMatrix}
\alias{contrastMatrix,DesignContrast-method}
\alias{contrastMatrix<-,DesignContrast-method}
\alias{contrastMatrix,MArrayLM-method}
\title{Extract the contrast matrix from an object}
\usage{
contrastMatrix(object)

\S4method{contrastMatrix}{DesignContrast}(object)

\S4method{contrastMatrix}{DesignContrast}(object) <- value

\S4method{contrastMatrix}{MArrayLM}(object)
}
\arguments{
\item{object}{An object, see supported methods below}

\item{value}{A contrast matrix}
}
\value{
A numeric contrast matrix
}
\description{
Extract the contrast matrix from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{contrastMatrix(DesignContrast)}: Return the contrast matrix from a DesignContrast
object

\item \code{contrastMatrix(DesignContrast) <- value}: Assign a contrast matrix to a \code{contrastContrast}
object

\item \code{contrastMatrix(MArrayLM)}: Extract contrast matrix from an object of MArrayLM

}}
