% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{designMatrix}
\alias{designMatrix}
\alias{designMatrix,DesignContrast-method}
\alias{designMatrix,MArrayLM-method}
\title{Extract the design matrix from an object}
\usage{
designMatrix(object)

\S4method{designMatrix}{DesignContrast}(object)

\S4method{designMatrix}{MArrayLM}(object)
}
\arguments{
\item{object}{An object, see supported methods below}
}
\description{
Extract the design matrix from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{DesignContrast}: Return the design matrix from a DesignContrast 
object

\item \code{MArrayLM}: Extract design matrix from an object of MArrayLM
}}

