% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{dispGroups}
\alias{dispGroups}
\alias{dispGroups,DesignContrast-method}
\title{Extract displayed group labels from an object}
\usage{
dispGroups(object)

\S4method{dispGroups}{DesignContrast}(object)
}
\arguments{
\item{object}{An object, see supported methods below}
}
\description{
Extract displayed group labels from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{dispGroups(DesignContrast)}: Return the sample groups from a DesignContrast object
, suing display labels

}}
