% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{eSetToLongTable}
\alias{eSetToLongTable}
\title{Transform eSet to long data.frame}
\usage{
eSetToLongTable(x, exprsFun = function(eset) Biobase::exprs(eset))
}
\arguments{
\item{x}{An \code{eSet} object}

\item{exprsFun}{A function to extract expression values, by default \code{exprs}

The function extracts gene expression, and return it in a long data.frame format with phenotypic data}
}
\description{
Transform eSet to long data.frame
}
\examples{
data(ribios.ExpressionSet, package="ribiosExpression")
exprsLongTbl <- eSetToLongTable(ribios.ExpressionSet)
seLongTbl <- eSetToLongTable(ribios.ExpressionSet, exprsFun=function(eset) assayData(eset)$se.exprs)
}
