% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{fixEmptyColumnName}
\alias{fixEmptyColumnName}
\title{Detect if any column has an empty string as name and fix}
\usage{
fixEmptyColumnName(df, prefix = "X")
}
\arguments{
\item{df}{A \code{data.frame}

If any column has an empty string as name, its replaced by the prefix appended by an index starting from 1}
}
\description{
Detect if any column has an empty string as name and fix
}
\examples{
testDf <- data.frame("Col1"=LETTERS[1:3], "Col2"=letters[2:4])
colnames(testDf) <- c("", "")
testDf
fixEmptyColumnName(testDf)
fixEmptyColumnName(testDf, prefix="fData")
}
