% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{nContrast}
\alias{nContrast}
\alias{nContrast,DesignContrast-method}
\title{Extract the number of contrasts from an object}
\usage{
nContrast(object)

\S4method{nContrast}{DesignContrast}(object)
}
\arguments{
\item{object}{An object, see supported methods below}
}
\description{
Extract the number of contrasts from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{nContrast(DesignContrast)}: Return the number of contrast in a DesignContras
object

}}
