% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{parseDesignContrastStr}
\alias{parseDesignContrastStr}
\title{Parse design and contrast from strings}
\usage{
parseDesignContrastStr(groupsStr, levelStr, dispLevelStr, contrastStr)
}
\arguments{
\item{groupsStr}{A factor vector indicating sample groups}

\item{levelStr}{Level strings}

\item{dispLevelStr}{Display level strings}

\item{contrastStr}{A vector of character strings indicating contrasts}
}
\value{
A DesignContrast object
}
\description{
Parse design and contrast from strings
}
