% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{plot.DesignContrast}
\alias{plot.DesignContrast}
\title{Plot a DesignContrast object}
\usage{
\method{plot}{DesignContrast}(x, y = NULL, name = NULL, ...)
}
\arguments{
\item{x}{A \code{DesignContrast} object}

\item{y}{NULL, ignored}

\item{name}{Name of the object, used in the title of the heatmaps}

\item{...}{Other parameters passed to \code{\link{biosHeatmap}}}
}
\value{
An invisible list of two elements, containing return values of \code{\link{biosHeatmap}} for design and contrast matrices, respectively.
}
\description{
The function plots two heatmaps, one of the design matrix and the other of the contrast matrix.
}
\examples{
myFac <- gl(3,3, labels=c("baseline", "treat1", "treat2"))
myDesign <- model.matrix(~myFac)
colnames(myDesign) <- c("baseline", "treat1", "treat2")
myContrast <- limma::makeContrasts(contrasts=c("treat1", "treat2"), levels=myDesign)
res1 <- DesignContrast(myDesign, myContrast, groups=myFac)
res2 <- DesignContrast(myDesign, myContrast, groups=myFac, dispLevels=c("C", "T1", "T2"))
plot(res1, name="Res 1")
plot(res2, name="Res 2")
}
