% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{plot.DesignContrast}
\alias{plot.DesignContrast}
\title{Plot a DesignContrast object with two heatmaps}
\usage{
\method{plot}{DesignContrast}(
  x,
  y = NULL,
  title = NULL,
  clusterDesign = FALSE,
  clusterSamples = FALSE,
  clusterContrasts = FALSE,
  designRange = NULL,
  contrastRange = NULL,
  designParams = NULL,
  contrastParams = NULL
)
}
\arguments{
\item{x}{A \code{DesignContrast} object}

\item{y}{NULL, ignored}

\item{title}{Title of the object, used in the title of the heatmaps}

\item{clusterDesign}{Logical, cluster rows of the design matrix}

\item{clusterSamples}{Logical, cluster columns of the design matrix}

\item{clusterContrasts}{Logical, cluster rows of the contrast matrix (notice that the contrast matrix required by limma is the transposed contrast matrix)}

\item{designRange}{NULL or a vector of length 2, giving range of the design matrix to be visualized. If NULL, the whole range is used.}

\item{contrastRange}{NULL or a vector of length 2, giving range of the contrast matrix to be visualized. If NULL, a symmetric range of the largest absolute value and its negate is used.}

\item{designParams}{Other parameters passed to \code{\link[ComplexHeatmap]{Heatmap} for the design matrix}}

\item{contrastParams}{Other parameters passed to \code{\link[ComplexHeatmap]{Heatmap} for the contrast matrix}}
}
\value{
Heatmap object
}
\description{
The function plots a ComplexHeatmap containing two matrices, one of the design matrix and the other of the contrast matrix.
}
\examples{
myFac <- gl(3,3, labels=c("baseline", "treat1", "treat2"))
myDesign <- model.matrix(~myFac)
colnames(myDesign) <- c("baseline", "treat1", "treat2")
myContrast <- limma::makeContrasts(contrasts=c("treat1", "treat2"), levels=myDesign)
res1 <- DesignContrast(myDesign, myContrast, groups=myFac)
res2 <- DesignContrast(myDesign, myContrast, groups=myFac, dispLevels=c("C", "T1", "T2"))
plot(res1, title="DesCon 1")
plot(res2, title="DesCon 1 (identical)")
plot(res2, title="DesCon 1 (identical)", designRange=c(-2,2), 
    contrastRange=c(-2,1),
    designParams=list(row_names_gp=gpar(fontsize=8)),
    contrastParams=list(column_names_gp=gpar(fontsize=12, color="red")))
}
