% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignContrast.R
\name{plot.DesignContrast}
\alias{plot.DesignContrast}
\title{Plot a DesignContrast object with two heatmaps}
\usage{
\method{plot}{DesignContrast}(
  x,
  y = NULL,
  title = NULL,
  clusterDesign = FALSE,
  clusterSamples = FALSE,
  clusterContrasts = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{DesignContrast} object}

\item{y}{NULL, ignored}

\item{title}{Title of the object, used in the title of the heatmaps}

\item{clusterDesign}{Logical, cluster rows of the design matrix}

\item{clusterSamples}{Logical, cluster columns of the design matrix}

\item{clusterContrasts}{Logical, cluster rows of the contrast matrix (notice that the contrast matrix required by limma is the transposed contrast matrix)}

\item{...}{Other parameters passed to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
Heatmap object
}
\description{
The function plots a ComplexHeatmap containing two matrices, one of the design matrix and the other of the contrast matrix.
}
\examples{
myFac <- gl(3,3, labels=c("baseline", "treat1", "treat2"))
myDesign <- model.matrix(~myFac)
colnames(myDesign) <- c("baseline", "treat1", "treat2")
myContrast <- limma::makeContrasts(contrasts=c("treat1", "treat2"), levels=myDesign)
res1 <- DesignContrast(myDesign, myContrast, groups=myFac)
res2 <- DesignContrast(myDesign, myContrast, groups=myFac, dispLevels=c("C", "T1", "T2"))
plot(res1, title="DesCon 1")
plot(res2, title="DesCon 1 (identical)")
}
