% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFullRank.R
\name{removeColRank}
\alias{removeColRank}
\title{Return rank of the matrix and the ranks of resulting matrices when each column is removed}
\usage{
removeColRank(matrix)
}
\arguments{
\item{matrix}{A numeric matrix}
}
\value{
A \code{data.frame} with \code{n+1} rows, where \code{n} is the column count of the input matrix
}
\description{
Return rank of the matrix and the ranks of resulting matrices when each column is removed
}
\examples{
myMat <- matrix(c(1,1,1, 0,1,1, 0,0,1, 1,0,0), ncol=4, byrow=FALSE)
removeColRank(myMat)
}
