% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosExpression-package.R
\docType{package}
\name{ribiosExpression-package}
\alias{ribiosExpression-package}
\title{Data structures and functions for gene expression analysis}
\description{
Provides data structures and functions for gene expression analysis
}
