% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosExpressionSet.R
\docType{data}
\name{ribiosExpressionSet}
\alias{ribiosExpressionSet}
\title{ribiosExpressionSet: An example of ExpressionSet with artificial expression data}
\description{
ribiosExpressionSet: An example of ExpressionSet with artificial expression data
}
\author{
Jitao David Zhang \email{jitao_david.zhang@roche.com}
}
\keyword{data}
