% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{rowscale.ExpressionSet}
\alias{rowscale.ExpressionSet}
\title{Perform row-wise scaling to an ExpressionSet object}
\usage{
\method{rowscale}{ExpressionSet}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{An ExpressionSet object.}

\item{center}{Logical, whether the mean values of rows should be set to zero.}

\item{scale}{Logical, whether the standard deviations of rows should be normalised to one.}
}
\description{
Perform row-wise scaling to an ExpressionSet object
}
