% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/io_gct_cls.R
\name{writeGct}
\alias{writeGct}
\alias{writeGct,matrix,ANY,ANY,ANY-method}
\alias{writeGct,eSet,ANY,ANY,ANY-method}
\alias{writeGct,matrix-method}
\alias{writeGct,eSet-method}
\title{Export matrix or eSet that can be coerced as one into gct/cls files}
\usage{
writeGct(obj, file, feat.name, feat.desc)

\S4method{writeGct}{matrix}(obj, file, feat.name, feat.desc)

\S4method{writeGct}{eSet}(obj, file, feat.name, feat.desc)
}
\arguments{
\item{obj}{The input object, see methods below for supported data types}

\item{file}{The output file}

\item{feat.name}{Specifying feature names}

\item{feat.desc}{Specifying feature descriptions}
}
\value{
Used for its side effect of writing files; returns invisibly.
}
\description{
Export matrix or eSet that can be coerced as one into gct/cls files
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Method for matrix as input, 
feta.name and feat.desc are passed to \code{write_gct}.

\item \code{eSet}: Use \code{eSet} as input. \code{feat.name} and \code{feat.desc} are 
variable (column) names in \code{fData}.
}}

\keyword{methods}
