% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tab.R
\name{writeSampleGroups}
\alias{writeSampleGroups}
\title{Write sample groups and group levels into plain text files}
\usage{
writeSampleGroups(sampleGroups, sampleGroups.file, sampleGroupLevels.file)
}
\arguments{
\item{sampleGroups}{Factor, encoding sample groups.}

\item{sampleGroups.file}{Character, file name where the information of sample groups is written to.}

\item{sampleGroupLevels.file}{Character, file name where the information of sample group levels is written to.

The function is used to export sample group and group level information for differential gene expression analysis.}
}
\description{
Write sample groups and group levels into plain text files
}
\examples{
writeSampleGroups(gl(3,4), stdout(), stdout())
}
